Sales Prediction/requirements.txt
phidata
openai
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq